LayoutTests/css3/flexbox/intrinsic-width-orthogonal-writing-mode.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
.vertical {
    writing-mode: vertical-lr;
}
</style>

<script src="../../resources/check-layout.js"></script>
<body onload="checkLayout('.inline-flexbox')">

<p>Tests that we calculate correct intrinsic widths when orthogonal flows are involved.</p>

<div class="inline-flexbox" data-expected-width="20">
    <div class="vertical" style="width: 20px; height: 50px;"></div>
</div>

<div class="inline-flexbox" data-expected-width="20">
    <div class="vertical" style="line-height: 20px;">Some text</div>
</div>

<div class="inline-flexbox" data-expected-width="20">
    <span class="vertical" data-expected-width="20" style="line-height: 20px;">Some text</span>
</div>
