Safari-609.1.11.1/LayoutTests/css3/flexbox/intrinsic-width-orthogonal-writing-mode-expected.txt

Tests that we calculate correct intrinsic widths when orthogonal flows are involved.

PASS
Some text
PASS
Some text
FAIL:
Expected 20 for width, but got 0. 

<div class="inline-flexbox" data-expected-width="20">
    <span class="vertical" data-expected-width="20" style="line-height: 20px;">Some text</span>
</div>
